# target/classes/static/openapi.yaml
openapi: 3.1.0
info:
  title: Book API
  version: 1.0.0
  description: API for managing books
servers:
  - url: http://localhost:8080
paths:
  /books:
    get:
      summary: Get all books
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookList'
    post:
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
            example:
              title: "The Pragmatic Programmer"
              author: "Andrew Hunt, David Thomas"
              price: 42.99
              CoverImageUrl: "https://m.media-amazon.com/images/I/41as+WafrFL._SX218_BO1,204,203,200_QL40_FMwebp_.jpg"
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
  /books/{id}:
    get:
      summary: Get a book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Book found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found
    put:
      summary: Update a book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Book updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
    delete:
      summary: Delete a book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Book deleted
components:
  schemas:
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "The Pragmatic Programmer"
        author:
          type: string
          example: "Andrew Hunt, David Thomas"
        price:
          type: number
          format: double
          example: 42.99
        CoverImageUrl:
          type: string
          example: "https://m.media-amazon.com/images/I/41as+WafrFL._SX218_BO1,204,203,200_QL40_FMwebp_.jpg"
    BookList:
      type: array
      items:
        $ref: '#/components/schemas/Book'
      example:
        - id: 1
          title: "The Pragmatic Programmer"
          author: "Andrew Hunt, David Thomas"
          price: 42.99
          CoverImageUrl: "https://m.media-amazon.com/images/I/41as+WafrFL._SX218_BO1,204,203,200_QL40_FMwebp_.jpg"
        - id: 2
          title: "Clean Code"
          author: "Robert C. Martin"
          price: 35.50
          CoverImageUrl: "https://m.media-amazon.com/images/I/41xShlnTZTL._SX374_BO1,204,203,200_.jpg"
        - id: 3
          title: "Effective Java"
          author: "Joshua Bloch"
          price: 50.00
          CoverImageUrl: "https://m.media-amazon.com/images/I/41bWcNdTG+L._SX396_BO1,204,203,200_.jpg"
        - id: 4
          title: "Head First Design Patterns"
          author: "Eric Freeman, Bert Bates"
          price: 45.99
          CoverImageUrl: "https://m.media-amazon.com/images/I/51k+eHh3y-L._SX430_BO1,204,203,200_.jpg"
        - id: 5
          title: "Introduction to Algorithms"
          author: "Thomas H. Cormen, Charles E. Leiserson, Ronald L. Rivest, Clifford Stein"
          price: 80.00
          CoverImageUrl: "https://m.media-amazon.com/images/I/41SNlZQ2HkL._SX396_BO1,204,203,200_.jpg"